Move (some) enums to separate project, #26
