+++ Camille.Enums/Camille.Enums.csproj
@@ -1,0 +1,73 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard1.1;net45;netcoreapp2.0;netstandard2.1;netcoreapp3.0;netcoreapp3.1</TargetFrameworks>
+    <LangVersion>8.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <PackageId>Camille.Enums</PackageId>
+    <RootNamespace>Camille.Enums</RootNamespace>
+    <Version>3.0.0</Version>
+    <Authors>MingweiSamuel</Authors>
+    <Company>MingweiSamuel</Company>
+    <PackageLicenseUrl>https://www.gnu.org/licenses/gpl-3.0.txt</PackageLicenseUrl>
+    <Copyright>LGPLv3 2019</Copyright>
+    <PackageProjectUrl>https://github.com/MingweiSamuel/Camille</PackageProjectUrl>
+    <PackageIconUrl>http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Camille.png</PackageIconUrl>
+    <RepositoryUrl>https://github.com/MingweiSamuel/Camille</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <Description>Enums for League of Legends and other Riot Games games.</Description>
+    <Summary>Enums for League of Legends and other Riot Games games.</Summary>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+  </PropertyGroup>
+  <PropertyGroup>
+    <GenDir>gen</GenDir>
+  </PropertyGroup>
+
+  <!-- Camille using Newtonsoft.Json. -->
+  <PropertyGroup Condition="'$(TargetFramework)' == 'netstandard1.1' Or '$(TargetFramework)' == 'net45' Or '$(TargetFramework)' == 'netcoreapp2.0' Or '$(TargetFramework)' == 'netcoreapp3.0'">
+    <DefineConstants>$(DefineConstants);USE_NEWTONSOFT</DefineConstants>
+  </PropertyGroup>
+  <!-- Camille using System.Text.Json. -->
+  <PropertyGroup Condition="'$(TargetFramework)' == 'netstandard2.1' Or '$(TargetFramework)' == 'netcoreapp3.1'">
+    <DefineConstants>$(DefineConstants);USE_SYSTEXTJSON</DefineConstants>
+  </PropertyGroup>
+
+  <!-- Newtonsoft & Nullable attributes from Nuget. -->
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.1' Or '$(TargetFramework)' == 'net45' Or '$(TargetFramework)' == 'netcoreapp2.0' Or '$(TargetFramework)' == 'netcoreapp3.0'">
+    <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
+    <PackageReference Include="Nullable" Version="1.1.0">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
+  <!-- System.Net.Http from Nuget. -->
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.1' Or '$(TargetFramework)' == 'net45'">
+    <PackageReference Include="System.Net.Http" Version="4.3.3" />
+  </ItemGroup>
+  <!-- System.Text.Json from Nuget. -->
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
+    <PackageReference Include="System.Text.Json" Version="4.6.0" />
+  </ItemGroup>
+  <!-- System.ComponentModel.DataAnnotations from Nuget. -->
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.1' Or '$(TargetFramework)' == 'net45' Or '$(TargetFramework)' == 'netstandard2.1'">
+    <PackageReference Include="System.ComponentModel.Annotations" Version="4.6.0" />
+  </ItemGroup>
+
+  <Target Name="DownloadTemplateData">
+    <DownloadFile SourceUrl="http://raw.communitydragon.org/pbe/plugins/rcp-be-lol-game-data/global/default/v1/champion-summary.json" DestinationFolder="$(GenDir)" DestinationFileName=".champion.json" ContinueOnError="true" />
+    <DownloadFile SourceUrl="http://static.developer.riotgames.com/docs/lol/seasons.json" DestinationFolder="$(GenDir)" DestinationFileName=".seasons.json" ContinueOnError="true" />
+    <DownloadFile SourceUrl="http://static.developer.riotgames.com/docs/lol/queues.json" DestinationFolder="$(GenDir)" DestinationFileName=".queues.json" ContinueOnError="true" />
+    <DownloadFile SourceUrl="http://static.developer.riotgames.com/docs/lol/maps.json" DestinationFolder="$(GenDir)" DestinationFileName=".maps.json" ContinueOnError="true" />
+    <DownloadFile SourceUrl="http://static.developer.riotgames.com/docs/lol/gameModes.json" DestinationFolder="$(GenDir)" DestinationFileName=".gameModes.json" ContinueOnError="true" />
+    <DownloadFile SourceUrl="http://static.developer.riotgames.com/docs/lol/gameTypes.json" DestinationFolder="$(GenDir)" DestinationFileName=".gameTypes.json" ContinueOnError="true" />
+  </Target>
+
+  <!-- On Windows DispatchToInnerBuilds is run due to multiple TargetFrameworks. -->
+  <Target Name="BuildTemplates" BeforeTargets="DispatchToInnerBuilds" DependsOnTargets="DownloadTemplateData">
+    <Exec WorkingDirectory="$(SolutionDir)" Command="node srcgen/index.js &quot;$(MSBuildProjectDirectory)&quot;" />
+  </Target>
+  <!-- On Unix/OSX DispatchToInnerBuilds is not run so we use GenerateTargetFrameworkMonikerAttribute. -->
+  <Target Name="BuildTemplatesUnix" BeforeTargets="GenerateTargetFrameworkMonikerAttribute" DependsOnTargets="BuildTemplates" Condition="$(OS) == 'Unix'">
+    <Message Text="Building templates on unix." />
+  </Target>
+</Project>
