+++ Camille.Enums/gen/.gitignore
