+++ Camille.Enums/gen/Champion.cs.dt
@@ -1,5 +1,5 @@
 ﻿{{
-    const champions = require('./.champion.json')
+    const champions = require(it.path + '/.champion.json')
         .filter(({ id }) => id > 0)
         .sortBy(({ name }) => name);
     const enumName = function(name) { return name.toUpperCase().replace(/[^A-Z]+/, '_'); };
@@ -10,7 +10,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Camille.RiotApi.Enums
+namespace Camille.Enums
 {
     /// <summary>
     /// Champion enum based on ChampionIds.
