+++ Camille.Enums/gen/GameMode.cs.dt
@@ -1,5 +1,5 @@
 {{
-    const queues = require('./.queues.json');
+    const gameModes = require(it.path + '/.gameModes.json');
     const enumName = function(name) { return name.toUpperCase().replace(/['-]/, '').replace(/[^A-Z0-9]+/g, '_'); };
 }}
 // This file is automatically generated.
@@ -8,21 +8,21 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Camille.RiotApi.Enums
+namespace Camille.Enums
 {
     /// <summary>
-    /// Queues enum based on queues.json.
+    /// GameModes enum based on gameModes.json.
     /// </summary>
 #if USE_SYSTEXTJSON
     [System.Text.Json.Serialization.JsonConverter(typeof(System.Text.Json.Serialization.JsonStringEnumConverter))]
 #endif
-    public enum Queue
+    public enum GameMode
     {
 {{
-    for (let { queueId, map, description, notes } of queues) {
+    for (let { gameMode, description } of gameModes) {
 }}
-        [Display(Name = "{{= map}}", Description = "{{= description}}")]
-        {{= enumName(notes !== null ? description === null ? map + "_DEPRECATED" : map + "_" + description + "_DEPRECATED" : map + "_" + description) }} = {{= queueId }},
+        [Display(Name = "{{= gameMode }}", Description = "{{= description }}")]
+        {{= enumName(gameMode) }},
 {{
     }
 }}
