+++ Camille.Enums/gen/GameType.cs.dt
@@ -1,5 +1,5 @@
 {{
-    const gameTypes = require('./.gameTypes.json');
+    const gameTypes = require(it.path + '/.gameTypes.json');
     const enumName = function(name) { return name.toUpperCase().replace(/['-]/, '').replace(/[^A-Z0-9]+/g, '_'); };
 }}
 // This file is automatically generated.
@@ -8,7 +8,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Camille.RiotApi.Enums
+namespace Camille.Enums
 {
     /// <summary>
     /// GameTypes enum based on gameTypes.json.
