+++ Camille.Enums/gen/Map.cs.dt
@@ -1,5 +1,5 @@
 {{
-    const maps = require('./.maps.json');
+    const maps = require(it.path + '/.maps.json');
     const enumName = function(name) { return name.toUpperCase().replace(/['-]/, '').replace(/[^A-Z0-9]+/g, '_'); };
 }}
 // This file is automatically generated.
@@ -8,7 +8,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Camille.RiotApi.Enums
+namespace Camille.Enums
 {
     /// <summary>
     /// Maps enum based on maps.json.
