+++ Camille.Enums/gen/Queue.cs.dt
@@ -1,5 +1,5 @@
 {{
-    const gameModes = require('./.gameModes.json');
+    const queues = require(it.path + '/.queues.json');
     const enumName = function(name) { return name.toUpperCase().replace(/['-]/, '').replace(/[^A-Z0-9]+/g, '_'); };
 }}
 // This file is automatically generated.
@@ -8,21 +8,21 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace Camille.RiotApi.Enums
+namespace Camille.Enums
 {
     /// <summary>
-    /// GameModes enum based on gameModes.json.
+    /// Queues enum based on queues.json.
     /// </summary>
 #if USE_SYSTEXTJSON
     [System.Text.Json.Serialization.JsonConverter(typeof(System.Text.Json.Serialization.JsonStringEnumConverter))]
 #endif
-    public enum GameMode
+    public enum Queue
     {
 {{
-    for (let { gameMode, description } of gameModes) {
+    for (let { queueId, map, description, notes } of queues) {
 }}
-        [Display(Name = "{{= gameMode }}", Description = "{{= description }}")]
-        {{= enumName(gameMode) }},
+        [Display(Name = "{{= map}}", Description = "{{= description}}")]
+        {{= enumName(notes !== null ? description === null ? map + "_DEPRECATED" : map + "_" + description + "_DEPRECATED" : map + "_" + description) }} = {{= queueId }},
 {{
     }
 }}
