+++ Camille.Enums/src/Locale.cs
@@ -1,7 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-
-namespace Camille.RiotApi.Enums
+﻿namespace Camille.Enums
 {
 #if USE_SYSTEXTJSON
   [System.Text.Json.Serialization.JsonConverter(typeof(System.Text.Json.Serialization.JsonStringEnumConverter))]
