+++ Camille.Enums/src/Tier.cs
