+++ Camille.RiotApi.Test/ApiChampionTest.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Camille.RiotApi.ChampionV3;
-using Camille.RiotApi.Enums;
+using Camille.Enums;
 
 namespace Camille.RiotApi.Test
 {
