+++ Camille.RiotApi.Test/ApiComboFeaturedGamesSummonerCurrentGameV4Test.cs
@@ -2,7 +2,7 @@
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Camille.RiotApi.Enums;
+using Camille.Enums;
 using Camille.RiotApi.SpectatorV4;
 
 namespace Camille.RiotApi.Test
