+++ Camille.RiotApi.Test/ApiLeagueV4Test.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Camille.RiotApi.Enums;
+using Camille.Enums;
 using Camille.RiotApi.LeagueV4;
 
 namespace Camille.RiotApi.Test
