+++ Camille.RiotApi/gen/DataClasses.cs.dt
@@ -1,5 +1,5 @@
 ﻿{{
-    const spec = require('./.spec.json');
+    const spec = require(it.path + '/.spec.json');
     const dotUtils = require('./dotUtils.js');
 }}
 // This file is automatically generated.
@@ -11,6 +11,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using Camille.Enums;
 using Camille.RiotApi.Enums;
 
 #if USE_NEWTONSOFT
