+++ Camille.RiotApi/gen/EndpointMethods.cs.dt
@@ -1,5 +1,5 @@
 ﻿{{
-    const spec = require('./.spec.json');
+    const spec = require(it.path + '/.spec.json');
     const dotUtils = require('./dotUtils.js');
 }}
 // This file is automatically generated.
@@ -13,6 +13,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Linq;
+using Camille.Enums;
 using Camille.RiotApi.Enums;
 
 namespace Camille.RiotApi
