+++ Camille.RiotApi/gen/RiotApiConfig.cs.dt
@@ -1,5 +1,5 @@
 ﻿{{
-    const spec = require('./configspec.json');
+    const spec = require(it.path + '/configspec.json');
     const dotUtils = require('./dotUtils.js');
 }}
 // This file is automatically generated.
