+++ Camille.RiotApi/gen/RiotApiEndpoints.cs.dt
@@ -1,5 +1,5 @@
 ﻿{{
-    const spec = require('./.spec.json');
+    const spec = require(it.path + '/.spec.json');
     const dotUtils = require('./dotUtils.js');
 }}
 // This file is automatically generated.
