+++ Camille.RiotApi/src/Enums/QueueType.cs
@@ -1,4 +1,6 @@
-﻿namespace Camille.RiotApi.Enums
+﻿using System;
+
+namespace Camille.RiotApi.Enums
 {
 #if USE_SYSTEXTJSON
   [System.Text.Json.Serialization.JsonConverter(typeof(System.Text.Json.Serialization.JsonStringEnumConverter))]
@@ -6,8 +8,9 @@
     public enum QueueType
     {
         RANKED_SOLO_5x5 = 420,
-        RANKED_TFT = 1100,
         RANKED_FLEX_SR = 440,
-        RANKED_FLEX_TT = 470
+        RANKED_FLEX_TT = 470,
+        [Obsolete("Replaced with dedicated TFT endpoints.")]
+        RANKED_TFT = 1100
     }
 }
