+++ Camille.RiotApi/src/Util/RegionalRequester.cs
@@ -5,7 +5,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Camille.RiotApi.Enums;
+using Camille.Enums;
 
 namespace Camille.RiotApi.Util
 {
