+++ Camille.RiotApi/src/Util/RequestManager.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Camille.RiotApi.Enums;
+using Camille.Enums;
 
 namespace Camille.RiotApi.Util
 {
