+++ Camille.sln
@@ -14,8 +14,21 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Camille.RiotApi", "Camille.RiotApi\Camille.RiotApi.csproj", "{7DB820F1-5E9B-4872-8B78-A1B799DC4B17}"
+	ProjectSection(ProjectDependencies) = postProject
+		{D0E87959-4FBD-4134-A399-3C29E9CE38A4} = {D0E87959-4FBD-4134-A399-3C29E9CE38A4}
+	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Camille.RiotApi.Test", "Camille.RiotApi.Test\Camille.RiotApi.Test.csproj", "{2798A2CC-31D0-490F-A257-A76E18414334}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Camille.Enums", "Camille.Enums\Camille.Enums.csproj", "{D0E87959-4FBD-4134-A399-3C29E9CE38A4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "srcgen", "srcgen", "{F671635E-4AA2-4A6E-90B2-4FD6C1D243BE}"
+	ProjectSection(SolutionItems) = preProject
+		srcgen\dotUtils.js = srcgen\dotUtils.js
+		srcgen\index.js = srcgen\index.js
+		srcgen\package.json = srcgen\package.json
+		srcgen\polyfill.js = srcgen\polyfill.js
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,9 +44,16 @@ Global
 		{2798A2CC-31D0-490F-A257-A76E18414334}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2798A2CC-31D0-490F-A257-A76E18414334}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2798A2CC-31D0-490F-A257-A76E18414334}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0E87959-4FBD-4134-A399-3C29E9CE38A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0E87959-4FBD-4134-A399-3C29E9CE38A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0E87959-4FBD-4134-A399-3C29E9CE38A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0E87959-4FBD-4134-A399-3C29E9CE38A4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F671635E-4AA2-4A6E-90B2-4FD6C1D243BE} = {69923FA5-A1C1-4BCB-BA3C-393B5548B6DF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A22938CF-A9E4-423B-84E6-0F0939DBE423}
